Split clean-scrub tests by concern
Move shared setup (the debug/production scenarios, conn_config method, populate
helper, workload constants) into helper_clean_scrub.py. Tests inherit
(CleanScrubBase, wttest.WiredTigerTestCase) so the mixin's conn_config method
takes precedence over the test base class's empty default.

  test_clean_scrub_eviction01      core (3 tests)
  test_clean_scrub_eviction02      accounting + configuration (5 tests)
  test_clean_scrub_eviction03      workload + concurrency (4 tests)
  test_clean_scrub_eviction_in_memory01
  test_clean_scrub_eviction_disagg01

26 cases total (12 methods x 2 scenarios + 1 in_memory + 1 disagg).
